refactor: extract metaverse room session
